Organizing the Quote stash
Organzing the quote stash by readin the file.csv, that acts as the dataset for the application, saving it in the object and calculating the polarity score uing the SentimentIntensityAnalyzer algorithm. After the saving the score in an object iterating through the quotes and saving the score with the quotes, by adding another column in the dataset. I add another column 'index' to create the sorted collection of the quote stash with indices.