Creating a dataset.
This method that I have created in this step is to create our stash of quotes, read from the quotes.csv file and calculated the polarity score. After calculating the polarity score of the quotes present in the stash, I have added another column in my stash to have these polarity scores for the quotes available to us in our quote stash.

Once added the sentiment_score column I have added another column to give the sorting number to my roqs of the stash in order to move through my stash up and  down the stash. this feature I will need later when I will create the method to generate quotes everytime the user asks for a newer quote according to his/her sentiments.