Requests from the user
In this file, I have created another method to handle the quotes' request from the user. Using the index we provided in the quotes stash, I am able to move up and down on the basis of the mood of the user. If the user selects the motivational quote, I move up from the cuurent index and down in the other case.